[R-7.2.3] uwb: update qm35 driver
Features:
* When the QM ROM code reports an error during a 2KB chunk transfer,
the driver now sends the same chunk up to 10 times, after waiting 65ms.
We are limiting the global number of attempts to 50.
* When the QM FW updater code reports an error during a 3KB chunk transfer,
the driver now sends the same chunk up to 3 times.
We are limiting the global number of attempts to 60.
* When the QM ROM code reports that the FW flashing has completed,
the driver now checks if QM has started or if QM is in ROM code.
If QM is in ROM code, the driver will restart the flashing procedure.
* We are limiting the global number of flashing attempts to 10.

Fixes:
* Reduce firmware flash time by about 2.5 seconds

Miscellaneous:
* FW updater code review follow up
* Address issues reported by static code analysis tool

Bug: 305587526
Bug: 305153260
Bug: 300954022
Bug: 305042154
Bug: 287112888
Bug: 289878908
Bug: 293241279
Bug: 294255598
Change-Id: I73f2f0ed277b7a65f681c4d4db690f37d7323dae
Signed-off-by: Neil Di Folco <neil.difolco@qorvo.com>